Move IdentityProviderConfiguration to identity-spi
