+++ core/identity-spi/src/main/java/software/amazon/awssdk/identity/spi/internal/DefaultIdentityProviderConfiguration.java
@@ -13,14 +13,14 @@
  * permissions and limitations under the License.
  */
 
-package software.amazon.awssdk.http.auth.spi.internal;
+package software.amazon.awssdk.identity.spi.internal;
 
 import java.util.HashMap;
 import java.util.Map;
 import software.amazon.awssdk.annotations.SdkInternalApi;
-import software.amazon.awssdk.http.auth.spi.IdentityProviderConfiguration;
 import software.amazon.awssdk.identity.spi.Identity;
 import software.amazon.awssdk.identity.spi.IdentityProvider;
+import software.amazon.awssdk.identity.spi.IdentityProviderConfiguration;
 import software.amazon.awssdk.utils.ToString;
 import software.amazon.awssdk.utils.Validate;
